Jan 2025 to Mar 2025 update.
This commit is the initial version 1.0 of the analyzer. Scheduled to be updated on the 24th of every month to include the update for the next month. This base version supports January, February and March analysis, Jan and Feb YTD analysis. More graphs and patterns coming soon as well as a comparison mode.